+++ Client/CaveGame/MCTilemap.h
@@ -47,13 +47,13 @@ public:
 	static constexpr int CHUNK_HEIGHT = 32;
 
 private:
-	int tileData[CHUNK_WIDTH][CHUNK_HEIGHT][CHUNK_WIDTH];
+	uint8_t tileData[CHUNK_WIDTH][CHUNK_HEIGHT][CHUNK_WIDTH];
 
 public:
 	MCTileChunk();
 
-	void SetTile(int x, int y, int z, int tile);
-	int GetTile(int x, int y, int z) const;
+	void SetTile(int x, int y, int z, uint8_t tile);
+	uint8_t GetTile(int x, int y, int z) const;
 };
 
 class MCTilemap
@@ -72,10 +72,10 @@ private:
 	std::vector<std::function<void(MCTileChunk*, int, int)>> notifyCallback;
 
 public:
-	void SetTile(int x, int y, int z, int tile, bool notify = false) noexcept;
-	void SetTile(const glm::ivec3& v, int tile, bool notify = false) noexcept;
-	int GetTile(int x, int y, int z) const noexcept;
-	int GetTile(const glm::ivec3& v) const noexcept;
+	void SetTile(int x, int y, int z, uint8_t tile, bool notify = false) noexcept;
+	void SetTile(const glm::ivec3& v, uint8_t tile, bool notify = false) noexcept;
+	uint8_t GetTile(int x, int y, int z) const noexcept;
+	uint8_t GetTile(const glm::ivec3& v) const noexcept;
 	MCTileChunk*const GetChunk(int x, int z)noexcept;
 	bool HandleCollision(const glm::vec3& pre_position, glm::vec3& position, glm::vec3& velocity)const noexcept;
 	RaycastResult RaycastTile(const glm::vec3& start_position, const glm::vec3& direction, float distance) const noexcept;
